Add eject alias as sometimes after using eject disc button another disc can't be inserted.
It seems that after successfully ejecting a disc that something
internally prevents another disc from being inserted. Not sure why this
happens. If anything the only problem I’ve ever had is not being able
to eject using the eject key but this command fixes that issue and
allows a new disc to be inserted.